requirements.txt
aiofiles
chromadb
edgartools
langchain-core
langchain-openai
langchain-text-splitters
matplotlib
neo4j
networkx
numpy
openai
pydantic
python-dotenv
rank-bm25
seaborn
tenacity
tiktoken
tqdm
